[breaking-change] Replace "contexts" with (design) "steps".
Instead of a fixed set of contexts the CLI now look at
the `[chipflow.steps]` section in the `chipflow.toml` file and loads
whichever classes are defined there.

Each class is responsible for building its own argument parser as
well as executing the action with its CLI. The default step
definitions are the same as the contexts they replace but the naming
is different and the downstream code needs to be updated.

This way steps can be easily customized both in ChipFlow code and in
customer code, including customizations that depend on the config
file, the platform, etc. Customers can also define their own steps
easily, or (potentially many) variations of built-in ones.