[breaking-change] Replace Sky130Platform with SiliconPlatform.
The new platform is not exactly compatible, but works in a similar
way. The customer needs to define a [chipflow.silicon.pads] section
in the `chipflow.toml` file for the platform to work.

The new platform uses different IO buffers and top interface and
requires backend work to fully integrate.

The HyperRAM provider isn't yet adjusted for the new platform.